Throw error in fields builder if form context is present
Didn't add the tests because not really sure how to test an error thrown
by a component without installing an external dependency. Although the
confidence for working of the code comes from the fact that other
tests are passing. That said, if a way is found for testing the above
mentioned scenario, then will definitely add the tests in the future.